Updated file to make betty databale

+++ 0x12-more_singly_linked_lists/0-print_listint.c
@@ -3,14 +3,14 @@
 #include "lists.h"
 
 /**
- * print_listin - Prints the elements of a list
+ * print_listint - Prints the elements of a list
  * @h: list to print
  *
  * Return: number of nodes
  */
 size_t print_listint(const listint_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	if (h == NULL)
 		return (0);
